refactor: 💡 updated folder names
