+++ src/components/NavbarComponent.tsx
@@ -1,6 +1,5 @@
 import {
   Flex,
-  useColorModeValue,
   Stack,
   IconButton,
   ButtonGroup,
@@ -11,7 +10,7 @@ import { CloseIcon } from '@chakra-ui/icons';
 import { appWindow } from '@tauri-apps/api/window';
 import { BiWindow } from 'react-icons/bi';
 import { MdOutlineMinimize } from 'react-icons/md';
-import AppSettings from './AppSetting/AppSettings';
+import AppSettings from './appsetting/AppSettings';
 import ThemeComponent from './ThemeComponent';
 import CommandHistory from './common/CommandHistory';
