+++ src/components/home/Info.tsx
