+++ src/components/home/ManjaroVersion.tsx
