+++ src/components/systemconfig/GpuInfoComponent.tsx
